Finally correctly finished GPU AI Node script
After all this time, finished writing the logic for the GPU AI Nodes. They needed to sum up the weights and inputs, then AFTER all that is done, apply a bias, then finally feed that result through an activation function of some sort (The magic logic that makes the AI work) in order to get an output. Still crashes, but the complicated part is done!